# services/openapi.yaml
openapi: 3.1.0
info:
  title: Effusion Labs Unified API
  version: 2.0.0
  description: |
    Unified API Gateway orchestrating all Effusion Labs microservices.
    
    Architecture:
    - BFF (Backend for Frontend) pattern
    - Event-driven realtime updates via WebSockets
    - Service mesh integration
    - MCP gateway proxy
    - Markdown conversion
    - Popmart recon automation
    - DayZ server management (future)
  contact:
    name: Effusion Labs
    url: https://effusionlabs.com

servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://localhost:18400
    description: Docker prod

tags:
  - name: health
    description: Health and monitoring endpoints
  - name: mcp
    description: MCP (Model Context Protocol) gateway integration
  - name: markdown
    description: URL to Markdown conversion
  - name: popmart
    description: Popmart recon automation
  - name: websocket
    description: Real-time WebSocket channels

paths:
  /:
    get:
      summary: API Root
      description: Discovery endpoint listing all available API routes
      tags: [health]
      responses:
        "200":
          description: API metadata and endpoint discovery
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  version:
                    type: string
                  endpoints:
                    type: object
                    additionalProperties:
                      type: string

  /health:
    get:
      summary: Orchestrator Health Check
      tags: [health]
      responses:
        "200":
          description: Orchestrator is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy]
                  timestamp:
                    type: string
                    format: date-time
                  services_monitored:
                    type: integer

  /services/health:
    get:
      summary: Aggregate Service Health
      description: Health status of all backend services
      tags: [health]
      responses:
        "200":
          description: List of service health statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceHealth"

  /mcp/servers:
    get:
      summary: List MCP Servers
      description: Get available MCP servers from gateway
      tags: [mcp]
      responses:
        "200":
          description: List of MCP servers
          content:
            application/json:
              schema:
                type: object
        "502":
          description: MCP Gateway unavailable

  /mcp/send:
    post:
      summary: Send MCP Request
      description: Forward request to specific MCP server
      tags: [mcp]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPRequest"
      responses:
        "200":
          description: MCP response
        "502":
          description: MCP Gateway error

  /markdown/convert:
    post:
      summary: Convert URL to Markdown
      description: Extract and convert webpage to markdown
      tags: [markdown]
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
      responses:
        "200":
          description: Markdown conversion successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  markdown:
                    type: string
        "502":
          description: Markdown Gateway error

  /popmart/recon:
    post:
      summary: Trigger Popmart Recon
      description: Start asynchronous popmart recon scan
      tags: [popmart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PopmartReconRequest"
      responses:
        "200":
          description: Recon task queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                  status:
                    type: string
                    enum: [queued]
                  message:
                    type: string

  /ws/{channel}:
    get:
      summary: WebSocket Channel
      description: Real-time bidirectional communication channel
      tags: [websocket]
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
            enum: [popmart, mcp, system, health]
      responses:
        "101":
          description: WebSocket connection established

  /events/stream:
    get:
      summary: Server-Sent Events
      description: SSE stream for real-time server events
      tags: [websocket]
      responses:
        "200":
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string

components:
  schemas:
    ServiceHealth:
      type: object
      required: [name, status, last_check]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [healthy, degraded, down]
        latency_ms:
          type: number
          nullable: true
        last_check:
          type: string
          format: date-time

    MCPRequest:
      type: object
      required: [server, payload]
      properties:
        server:
          type: string
          description: Name of MCP server
        payload:
          type: object
          description: JSON-RPC request payload

    PopmartReconRequest:
      type: object
      properties:
        id_start:
          type: integer
          minimum: 1
          default: 3450
        id_end:
          type: integer
          maximum: 10000
          default: 3650
        scan_collections:
          type: boolean
          default: false
        scan_brandip:
          type: boolean
          default: false
        scan_popnow:
          type: boolean
          default: false
        probe_cdn:
          type: boolean
          default: false